Use static_cast to avoid comparison of signed and unsigned integer warning in util.h.
git-svn-id: https://gflags.googlecode.com/svn/trunk@81 6586e3c6-dcc4-952a-343f-ff74eb82781d

+++ src/util.h
@@ -264,7 +264,7 @@ inline void InternalStringPrintf(std::string* output, const char* format,
   int bytes_written = vsnprintf(space, sizeof(space), format, backup_ap);
   va_end(backup_ap);
 
-  if ((bytes_written >= 0) && (bytes_written < sizeof(space))) {
+  if ((bytes_written >= 0) && (static_cast<size_t>(bytes_written) < sizeof(space))) {
     output->append(space, bytes_written);
     return;
   }
